@@ -1,56 +1,83 @@
 source 'https://rubygems.org'
-ruby "2.4.10"
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-gem 'rails', '5.2.4.4'
+ruby '2.6.6'
 
+gem 'rails', '~> 6.1.1'
+
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'dotenv-rails', '~> 2.7', '>= 2.7.4'
-gem 'pg'#, group: 'production'
-#gem 'mysql2', group: 'development'
-gem 'bootsnap', '~> 1.4', '>= 1.4.4'
-gem 'listen', '~> 3.1', '>= 3.1.5'
-
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
-
-# Gems used only for assets and not required
-# in production environments by default.
-#group :assets do
-  #gem 'sass-rails',   '~> 3.2.3'
-  #gem 'coffee-rails', '~> 3.2.1'
-
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer'
+gem 'pg'
 
 gem 'uglifier', '>= 1.0.3'
-#end
+# Use sqlite3 as the database for Active Record
+gem 'sqlite3', '~> 1.4'
+# Use Puma as the app server
+gem 'puma', '~> 5.0'
+# Use SCSS for stylesheets
+gem 'sass-rails', '>= 6'
+# Transpile app-like JavaScript. Read more: https://github.com/rails/webpacker
+gem 'webpacker', '~> 5.0'
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+gem 'turbolinks', '~> 5'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.7'
+# Use Redis adapter to run Action Cable in production
+# gem 'redis', '~> 4.0'
+# Use Active Model has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use Active Storage variant
+# gem 'image_processing', '~> 1.2'
 gem 'pry', '~> 0.13.1'
-gem 'byebug', '~>11.0.1'
-gem 'puma', '~> 4.0.1'
+
 gem 'twitter', '~> 6.2.0'
 gem 'sentry-raven', '~> 2.11'
-#gem 'jquery-rails'
-gem "devise", "~> 4.6.2"
+
+gem "devise", "~> 4.7.3"
 gem "omniauth", "~>1.9.0"
 gem "omniauth-twitter", "~> 1.4.0"
 gem "haml-rails", "~> 2.0.1"
-gem "twitter-bootstrap-rails", "~> 4.0.0"
+gem "twitter-bootstrap-rails", "~> 3.2.2"
 gem "redcarpet", "~> 1.17.2"
-gem "nokogiri", "~> 1.10.3"
+gem "nokogiri", "~> 1.10.10"
 gem "pygmentize", "~> 0.0.3"
-#gem "exception_notification", "~> 2.5.2"
+
 gem 'newrelic_rpm'
-gem "activeadmin", "~> 2.2.0"
-#gem "meta_search",    '>= 1.1.0.pre'
-#gem "therubyracer", "~> 0.11.2"
+gem "activeadmin", "~> 2.9.0"
+# Reduces boot times through caching; required in config/boot.rb
+gem 'bootsnap', '>= 1.4.4', require: false
+
 group :development, :test do
+  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
   gem 'rspec-rails'
+
+end
+
+group :development do
+  # Access an interactive console on exception pages or by calling 'console' anywhere in the code.
+  gem 'web-console', '>= 4.1.0'
+  # Display performance information such as SQL time and flame graphs for each request in your browser.
+  # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
+  gem 'rack-mini-profiler', '~> 2.0'
+  gem 'listen', '~> 3.3'
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
 end
 
 group :test do
+  # Adds support for Capybara system testing and selenium driver
+  gem 'capybara', '>= 3.26'
+  gem 'selenium-webdriver'
+  # Easy installation and use of web drivers to run system tests with browsers
+  gem 'webdrivers'
   gem 'database_cleaner'
   gem 'factory_bot_rails'
   gem 'faker'
   gem 'shoulda-matchers'
 end
 
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
 gem 'simplecov', require: false, group: :test
